agent-brain/requirements.txt
anthropic>=0.42.0
ddgs>=9.0.0
python-dotenv>=1.0.0
